requirements.txt
anthropic
streamlit
yfinance
pandas
plotly
python-dotenv
